Deleted two pages and turned two components into those two pages
